add files to the commit
